requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pytest==7.4.3
httpx==0.25.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
slowapi==0.1.9
prometheus-client==0.19.0
structlog==23.2.0

